api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx
psycopg[binary]      # si tu te connectes à Postgres en prod
pypdf>=4.0.0
python-multipart>=0.0.9
pydantic>=2.8
prometheus-client>=0.20.0
slowapi>=0.1.9
requests>=2.32.0
# si tu utilises PyPDF2 plutôt que pypdf:
# PyPDF2>=3.0.0
